fix(client-core): SignalService FileTransferReceived long to int
Interface was changed to int but SignalService field was still
long. CS0738 build error.

+++ src/PcBeaconAgent.Client.Core/Services/SignalService.cs
@@ -17,7 +17,7 @@ public class SignalService(ILogger<SignalService> mLogger, IPreferencesService m
 {
     public event Action<string, bool>? DeviceStatusChanged;
     public event Action<string, string, string>? TextTransferReceived;
-    public event Action<string, string, long, string, string>? FileTransferReceived;
+    public event Action<string, string, int, string, string>? FileTransferReceived;
 
     public string ClientMachineName { get; set; } = string.Empty;
